zmiana w plikach
