--- ProgNar/raport.txt
@@ -1,20 +1,0 @@
-=== Raport wyceny narzędzi skrawających ===
-
-Narzędzie 1:
-Typ frezu: Frez prosty
-Ilość ostrzy: 2-4
-Zakres średnicy: 6.1 - 10.0 mm
-Ilość sztuk: 10
-Cena jednostkowa: 41.70 PLN
-Cena całkowita: 417.00 PLN
-
-Narzędzie 2:
-Typ frezu: Frez promieniowy
-Ilość ostrzy: >4
-Zakres średnicy: 16.1 - 20.0 mm
-Ilość sztuk: 10
-Cena jednostkowa: 106.90 PLN
-Cena całkowita: 1069.00 PLN
-
-Suma za wszystkie pozycje: 1486.00 PLN
-====================================
